+++ src/components/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../store';
+import store from '../store/store';
 import Navigation from './Navigation';
-import Home from './Home';
+import Home from '../views/Home';
 
 const App = () => (
   <Provider store={store}>
