+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './stylesheet/index.css';
+import './assets/stylesheet/index.css';
 import App from './components/App';
 
 ReactDOM.render(
