+++ src/store/store.js
@@ -1,6 +1,6 @@
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers/index';
+import rootReducer from './reducers';
 
 const initialState = {};
 const middleware = [thunk];
